requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.3
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloud-init==22.3.4
cloudpickle==2.2.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
cryptography==2.8
cycler==0.11.0
databricks-cli==0.17.7
dbus-python==1.2.16
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
docker==6.1.3
docopt==0.6.2
dominate==2.7.0
entrypoints==0.3
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.10.7
findspark==2.0.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
fonttools==4.38.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
importlib-resources==5.12.0
incremental==16.10.1
ipykernel==6.17.0
ipython==8.6.0
ipython_genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaggle==1.5.12
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mlflow==2.4.1
more-itertools==4.2.0
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.1.0
openai==0.28.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2023.4.20
pipreqs==0.4.13
platformdirs==3.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.23.3
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycosat==0.6.3
pycparser==2.21
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
pyspark==3.4.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python-slugify==6.1.2
pytz==2022.6
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
querystring-parser==1.2.4
requests==2.31.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setuptools-scm==7.0.5
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sos==4.4
soupsieve==2.3.2.post1
SQLAlchemy==2.0.17
sqlparse==0.4.4
ssh-import-id==5.10
stack-data==0.6.0
systemd-python==234
tabulate==0.9.0
terminado==0.17.0
testresources==2.0.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.5.0
Twisted==18.9.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.16
virtualenv==20.21.0
virtualenv-clone==0.5.7
visitor==0.1.3
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
WTForms==3.0.1
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zope.interface==4.7.1
